requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.1
mypy==0.982
mypy-extensions==0.4.3
numpy==1.23.4
pandas==1.5.1
Pillow==9.3.0
python-dateutil==2.8.2
pytz==2022.6
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
tomli==2.0.1
typing_extensions==4.4.0
Werkzeug==2.2.2
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.3
Flask-Cors==3.0.10
keras==2.8.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-estimator==2.8.0
tensorflow-io-gcs-filesystem==0.31.0
urllib3==1.26.14
mysqlclient==2.1.1
protobuf==3.20.0
mysql-connector-python==8.0.29